+++ terrain/astar.h
@@ -5,7 +5,7 @@
 #include <queue>
 #include "graph.h"
 
-std::vector<Node> astarSearch(Graph *graph, std::unordered_map<int,int> heuristics, int start, int end);
+std::vector<Node> astarSearch(Graph *graph, std::unordered_map<int,float> heuristics, int start, int end);
 
 bool addToOpen(std::vector<Node> open, Node neighbor);
 #endif
